Enum timely::communication::allocator::generic::GenericBuilder

pub enum GenericBuilder {
    Thread(Thread),
    Process(Process),
    ProcessBinary(ProcessBuilder),
    ZeroCopy(TcpBuilder<Process>),
}

Enumerations of constructable implementors of Allocate.

The builder variants are meant to be Send, so that they can be moved across threads, whereas the allocator they construct may not. As an example, the ProcessBinary type contains Rc wrapped state, and so cannot itself be moved across threads.

Variants
Thread(Thread)

Builder for Thread allocator.

Process(Process)

Builder for Process allocator.

ProcessBinary(ProcessBuilder)

Builder for ProcessBinary allocator.

ZeroCopy(TcpBuilder<Process>)

Builder for ZeroCopy allocator.