Trait timely::communication::allocator::Allocate

pub trait Allocate {
    fn index(&self) -> usize;

    fn peers(&self) -> usize;

    fn allocate<T>(
        &mut self
    ) -> (Vec<Box<Push<Message<T>> + 'static>>, Box<Pull<Message<T>> + 'static>, Option<usize>)
    where
        T: Data;

    fn pre_work(&mut self) { ... }

    fn post_work(&mut self) { ... }
}

A type capable of allocating channels.

There is some feature creep, in that this contains several convenience methods about the nature of the allocated channels, and maintenance methods to ensure that they move records around.

Required Methods

fn index(&self) -> usize

The index of the worker out of (0..self.peers()).

fn peers(&self) -> usize

The number of workers.

fn allocate<T>(
&mut self
) -> (Vec<Box<Push<Message<T>> + 'static>>, Box<Pull<Message<T>> + 'static>, Option<usize>) where
T: Data,

Constructs several send endpoints and one receive endpoint.

Provided Methods

fn pre_work(&mut self)

Work performed before scheduling dataflows.

fn post_work(&mut self)

Work performed after scheduling dataflows.