Trait timely::communication::allocator::AllocateBuilder

pub trait AllocateBuilder: Send {
    type Allocator: Allocate;
    fn build(self) -> Self::Allocator;
}

A proto-allocator, which implements Send and can be completed with build.

This trait exists because some allocators contain non-Send elements, like Rc wrappers for shared state. As such, what we actually need to create to initialize a computation are builders, which we can then spawn in new threads each of which then construct their actual allocator.

Associated Types

type Allocator: Allocate

The type of built allocator.

Loading content...
Required methods

fn build(self) -> Self::Allocator

Builds allocator, consumes self.

Loading content...