Module timely::dataflow::channels

Structured communication between timely dataflow operators.

Reexports

pub use self::message::Message;

pub use self::message::Content;

Modules

message
Types relating to batching of timestamped records. Core type for communicating a collection of D: Data records.

pact
Parallelization contracts, describing how data must be exchanged between operators. Parallelization contracts, describing requirements for data movement along dataflow edges.

pullers
A collection of types that may be pulled from.

pushers
A collection of types that may be pushed at.