Module timely::dataflow::channels::message

Types relating to batching of timestamped records. Core type for communicating a collection of D: Data records.

Message<D> is meant to be treated as a Vec<D>, with the caveat that it may wrap either typed Vec<D> data or binary Vec<u8> data that have not yet been deserialized. The type implements Deref and DerefMut with Target = Vec<D>, whose implementations accommodate the possibly serialized representation.

Structs

Message
A serializable representation of timestamped data.

Enums

Content
A batch of data, represented either as serialized bytes or typed Rust objects.