Struct timely::dataflow::channels::pushers::buffer::Session

pub struct Session<'a, T, D, P: Push<(T, Content<D>)> + 'a> where
    T: Eq + Clone + 'a,
    D: 'a,  { /* fields omitted */ }

An output session for sending records at a specified time.

The Session struct provides the user-facing interface to an operator output, namely the Buffer type. A Session wraps a session of output at a specified time, and avoids what would otherwise be a constant cost of checking timestamp equality.

Methods

impl<'a, T, D, P: Push<(T, Content<D>)> + 'a> Session<'a, T, D, P> where
T: Eq + Clone + 'a,
D: 'a,

pub fn give(&mut self, data: D)

Provides one record at the time specified by the Session.

pub fn give_iterator<I: Iterator<Item = D>>(&mut self, iter: I)

Provides an iterator of records at the time specified by the Session.

pub fn give_content(&mut self, message: &mut Content<D>)

Provides a fully formed Content<D> message for senders which can use this type.

The Content type is the backing memory for communication in timely, and it can often be more efficient to re-use this memory rather than have timely allocate new backing memory.