Module timely::dataflow::operators

Extension traits for Stream implementing various operators.

A collection of functions taking typed Stream objects as input and producing new Stream objects as output. Many of the operators provide simple, composable functionality. Some of the operators are more complicated, for use with advanced timely dataflow features.

The Operator trait provides general operators whose behavior can be supplied using closures accepting input and output handles. Most of the operators in this module are defined using these two general operators.

Re-exports

pub use self::enterleave::Enter;

pub use self::enterleave::EnterAt;

pub use self::enterleave::Leave;

pub use self::input::Input;

pub use self::unordered_input::UnorderedInput;

pub use self::feedback::LoopVariable;

pub use self::feedback::ConnectLoop;

pub use self::concat::Concat;

pub use self::concat::Concatenate;

pub use self::partition::Partition;

pub use self::map::Map;

pub use self::inspect::Inspect;

pub use self::filter::Filter;

pub use self::delay::Delay;

pub use self::exchange::Exchange;

pub use self::broadcast::Broadcast;

pub use self::probe::Probe;

pub use self::to_stream::ToStream;

pub use self::capture::Capture;

pub use self::branch::Branch;

pub use self::branch::BranchWhen;

pub use self::generic::Operator;

pub use self::reclock::Reclock;

pub use self::count::Accumulate;

Modules

aggregation
Aggregation operators of various flavors

branch
Operators that separate one stream into two streams based on some condition

broadcast
Broadcast records to all workers.

capture
Operators to capture and replay timely dataflow streams.

concat
Merges the contents of multiple streams.

count
Counts the number of records at each time.

delay
Operators acting on timestamps to logically delay records

enterleave
Extension traits to move a Stream between an outer Scope and inner Scope.

exchange
Exchange records between workers.

feedback
Create cycles in a timely dataflow graph.

filter
Filters a stream by a predicate.

flow_controlled
Methods to construct flow-controlled sources.

generic
Generic operators defined by user-provided closures.

input
Create new Streams connected to external inputs.

inspect
Extension trait and implementation for observing and action on streamed data.

map
Extension methods for Stream based on record-by-record transformation.

partition
Partition a stream of records into multiple streams.

probe
Monitor progress at a Stream.

reclock
Extension methods for Stream based on record-by-record transformation.

to_stream
Conversion to the Stream type from iterators.

unordered_input
Create new Streams connected to external inputs.

Structs

Capability
The capability to send data with a certain timestamp on a dataflow edge.

CapabilityRef
An unowned capability, which can be used but not retained.

CapabilitySet
A set of capabilities, for possibly incomparable times.

FrontierNotificator
Tracks requests for notification and delivers available notifications.

Notificator
Tracks requests for notification and delivers available notifications.