Struct timely::dataflow::operators::capture::event::link::EventLink

pub struct EventLink<T, D> {
    pub event: Option<Event<T, D>>,
    pub next: RefCell<Option<Rc<EventLink<T, D>>>>,
}

A linked list of Event<T, D>.

Fields
event: Option<Event<T, D>>

An event, if one exists.

An event might not exist, if either we want to insert a None and have the output iterator pause, or in the case of the very first linked list element, which has no event when constructed.

next: RefCell<Option<Rc<EventLink<T, D>>>>

The next event, if it exists.

Methods

impl<T, D> EventLink<T, D>

pub fn new() -> EventLink<T, D>

Allocates a new EventLink.