Trait timely::dataflow::operators::capture::event::EventIterator

pub trait EventIterator<T, D> {
    fn next(&mut self) -> Option<&Event<T, D>>;
}

Iterates over contained Event<T, D>.

The EventIterator trait describes types that can iterate over references to events, and which can be used to replay a stream into a new timely dataflow computation.

This method is not simply an iterator because of the lifetime in the result.

Required methods

fn next(&mut self) -> Option<&Event<T, D>>

Iterates over references to Event<T, D> elements.

Loading content...