Trait timely::dataflow::operators::capture::replay::Replay

pub trait Replay<T: Timestamp, D: Data> {
    fn replay_into<S: Scope<Timestamp = T>>(self, scope: &mut S) -> Stream<S, D>;
}

Replay a capture stream into a scope with the same timestamp.

Required methods

fn replay_into<S: Scope<Timestamp = T>>(self, scope: &mut S) -> Stream<S, D>

Replays self into the provided scope, as a Stream<S, D>.

Loading content...