Trait timely::dataflow::operators::enterleave::Enter

pub trait Enter<G: Scope, T: Timestamp, D: Data> {
    fn enter<'a>(&self, _: &Child<'a, G, T>) -> Stream<Child<'a, G, T>, D>;
}

Extension trait to move a Stream into a child of its current Scope.

Required Methods

fn enter<'a>(&self, _: &Child<'a, G, T>) -> Stream<Child<'a, G, T>, D>

Moves the Stream argument into a child of its current Scope.

Examples

use timely::dataflow::scopes::Scope;
use timely::dataflow::operators::{Enter, Leave, ToStream};

timely::example(|outer| {
    let stream = (0..9).to_stream(outer);
    let output = outer.scoped::<u32,_,_>(|inner| {
        stream.enter(inner).leave()
    });
});