Module timely::dataflow::operators::feedback

Create cycles in a timely dataflow graph.

Structs

Handle
A handle used to bind the source of a loop variable.

Observer

Traits

ConnectLoop
Connect a Stream to the input of a loop variable.

LoopVariable
Creates a Stream and a Handle to later bind the source of that Stream.