Struct timely::dataflow::operators::flow_controlled::IteratorSourceInput

pub struct IteratorSourceInput<T: Clone, D: Data, DI: IntoIterator<Item = D>, I: IntoIterator<Item = (T, DI)>> {
    pub lower_bound: T,
    pub data: I,
    pub target: T,
}

Output of the input reading function for iterator_source.

Fields
lower_bound: T

Lower bound on timestamps that can be emitted by this input in the future.

data: I

Any T: IntoIterator of new input data in the form (time, data): time must be monotonically increasing.

target: T

A timestamp that represents the frontier that the probe should have reached before the function is invoked again to ingest additional input.