Module timely::dataflow::operators::generic

Generic operators defined by user-provided closures.

Reexports

pub use self::unary::Unary;

pub use self::binary::Binary;

pub use self::operator::Operator;

pub use self::operator::source;

Modules

binary
Methods to construct generic streaming and blocking binary operators.

builder_raw
Types to build operators with general shapes.

builder_rc
Types to build operators with general shapes.

operator
Methods to construct generic streaming and blocking unary operators.

unary
Methods to construct generic streaming and blocking unary operators.

Structs

FrontierNotificator
Tracks requests for notification and delivers available notifications.

FrontieredInputHandle
Handle to an operator's input stream and frontier.

InputHandle
Handle to an operator's input stream.

Notificator
Tracks requests for notification and delivers available notifications.

OutputHandle
Handle to an operator's output stream.