Module timely::dataflow::operators::generic::builder_raw

Types to build operators with general shapes.

These types expose some raw timely interfaces, and while public so that others can build on them, they require some sophistication to use correctly. I recommend checking out builder_rc.rs for an interface that is intentionally harder to mis-use.

Structs

OperatorBuilder
Builds operators with generic shape.

OperatorShape
Contains type-free information about the operator properties.