Struct timely::dataflow::operators::generic::builder_raw::OperatorBuilder

pub struct OperatorBuilder<G: Scope> { /* fields omitted */ }

Builds operators with generic shape.

Methods

impl<G: Scope> OperatorBuilder<G>

pub fn new(name: String, scope: G) -> Self

Allocates a new generic operator builder from its containing scope.

pub fn index(&self) -> usize

The operator's index

pub fn shape(&self) -> &OperatorShape

Return a reference to the operator's shape

pub fn set_notify(&mut self, notify: bool)

Indicates whether the operator requires frontier information.

pub fn new_input<D: Data, P>(
&mut self,
stream: &Stream<G, D>,
pact: P
) -> P::Puller where
P: ParallelizationContract<G::Timestamp, D>,

Adds a new input to a generic operator builder, returning the Pull implementor to use.

pub fn new_input_connection<D: Data, P>(
&mut self,
stream: &Stream<G, D>,
pact: P,
connection: Vec<Antichain<<G::Timestamp as Timestamp>::Summary>>
) -> P::Puller where
P: ParallelizationContract<G::Timestamp, D>,

Adds a new input to a generic operator builder, returning the Pull implementor to use.

pub fn new_output<D: Data>(&mut self) -> (Tee<G::Timestamp, D>, Stream<G, D>)

Adds a new input to a generic operator builder, returning the Push implementor to use.

pub fn new_output_connection<D: Data>(
&mut self,
connection: Vec<Antichain<<G::Timestamp as Timestamp>::Summary>>
) -> (Tee<G::Timestamp, D>, Stream<G, D>)

Adds a new input to a generic operator builder, returning the Push implementor to use.

pub fn build<PEP, PIP>(self, push_external: PEP, pull_internal: PIP) where
PEP: FnMut(&mut [ChangeBatch<G::Timestamp>]) + 'static,
PIP: FnMut(&mut [ChangeBatch<G::Timestamp>], &mut [ChangeBatch<G::Timestamp>], &mut [ChangeBatch<G::Timestamp>]) -> bool + 'static,

Creates an operator implementation from supplied logic constructor.