Struct timely::dataflow::operators::generic::builder_rc::OperatorBuilder

pub struct OperatorBuilder<G: Scope> { /* fields omitted */ }

Builds operators with generic shape.

Methods

impl<G: Scope> OperatorBuilder<G>

pub fn new(name: String, scope: G) -> Self

Allocates a new generic operator builder from its containing scope.

pub fn set_notify(&mut self, notify: bool)

Indicates whether the operator requires frontier information.

pub fn new_input<D: Data, P>(
&mut self,
stream: &Stream<G, D>,
pact: P
) -> InputHandle<G::Timestamp, D, P::Puller> where
P: ParallelizationContract<G::Timestamp, D>,

Adds a new input to a generic operator builder, returning the Pull implementor to use.

pub fn new_input_connection<D: Data, P>(
&mut self,
stream: &Stream<G, D>,
pact: P,
connection: Vec<Antichain<<G::Timestamp as Timestamp>::Summary>>
) -> InputHandle<G::Timestamp, D, P::Puller> where
P: ParallelizationContract<G::Timestamp, D>,

Adds a new input with connection information to a generic operator builder, returning the Pull implementor to use.

pub fn new_output<D: Data>(
&mut self
) -> (OutputWrapper<G::Timestamp, D, Tee<G::Timestamp, D>>, Stream<G, D>)

Adds a new output to a generic operator builder, returning the Pull implementor to use.

pub fn new_output_connection<D: Data>(
&mut self,
connection: Vec<Antichain<<G::Timestamp as Timestamp>::Summary>>
) -> (OutputWrapper<G::Timestamp, D, Tee<G::Timestamp, D>>, Stream<G, D>)

Adds a new output with connection information to a generic operator builder, returning the Pull implementor to use.

pub fn build<B, L>(self, constructor: B) where
B: FnOnce(Capability<G::Timestamp>) -> L,
L: FnMut(&[MutableAntichain<G::Timestamp>]) + 'static,

Creates an operator implementation from supplied logic constructor.