Module timely::dataflow::operators::input

Create new Streams connected to external inputs.

Structs

Handle
A handle to an input Stream, used to introduce data to a timely dataflow computation.

Traits

Input
Create a new Stream and Handle through which to supply input.