Struct timely::dataflow::operators::Capability

pub struct Capability<T: Timestamp> { /* fields omitted */ }

The capability to send data with a certain timestamp on a dataflow edge.

Capabilities are used by timely dataflow's progress tracking machinery to restrict and track when user code retains the ability to send messages on dataflow edges. All capabilities are constructed by the system, and should eventually be dropped by the user. Failure to drop a capability (for whatever reason) will cause timely dataflow's progress tracking to stall.

Methods

impl<T: Timestamp> Capability<T>

Important traits for &'a mut I
impl<'a, I> Iterator for &'a mut I where
I: Iterator + ?Sized, type Item = <I as Iterator>::Item;impl<'a, R> Read for &'a mut R where
R: Read + ?Sized, impl<'a, W> Write for &'a mut W where
W: Write + ?Sized,

pub fn time(&self) -> &T

The timestamp associated with this capability.

pub fn delayed(&self, new_time: &T) -> Capability<T>

Makes a new capability for a timestamp new_time greater or equal to the timestamp of the source capability (self).

This method panics if self.time is not less or equal to new_time.

pub fn downgrade(&mut self, new_time: &T)

Downgrades the capability to one corresponding to new_time.

This method panics if self.time is not less or equal to new_time.