Struct timely::dataflow::operators::CapabilitySet

pub struct CapabilitySet<T: Timestamp> { /* fields omitted */ }

A set of capabilities, for possibly incomparable times.

Methods

impl<T: Timestamp> CapabilitySet<T>

pub fn new() -> Self

Allocates an empty capability set.

pub fn insert(&mut self, capability: Capability<T>)

Inserts capability into the set, discarding redundant capabilities.

pub fn delayed(&self, time: &T) -> Capability<T>

Creates a new capability to send data at time.

This method panics if there does not exist a capability in self.elements less or equal to time.

pub fn downgrade(&mut self, frontier: &[T])

Downgrades the set of capabilities to correspond with the times in frontier.

This method panics if any element of frontier is not greater or equal to some element of self.elements.