Struct timely::dataflow::operators::unordered_input::UnorderedHandle

pub struct UnorderedHandle<T: Timestamp, D: Data> { /* fields omitted */ }

A handle to an input Stream, used to introduce data to a timely dataflow computation.

Methods

impl<T: Timestamp, D: Data> UnorderedHandle<T, D>

pub fn session<'b>(
&'b mut self,
cap: Capability<T>
) -> AutoflushSession<'b, T, D, PushCounter<T, D, Tee<T, D>>>

Allocates a new automatically flushing session based on the supplied capability.