Module timely::dataflow::scopes

Hierarchical organization of timely dataflow graphs.

Reexports

pub use self::child::Child;

pub use self::root::Root;

Modules

child
A child dataflow scope, used to build nested dataflow scopes.

root
The root scope of all timely dataflow computations.

Traits

Scope
The fundamental operations required to add and connect operators in a timely dataflow graph.

ScopeParent
The information a child scope needs from its parent.