Struct timely::dataflow::scopes::root::Root

pub struct Root<A: Allocate> { /* fields omitted */ }

A Root is the entry point to a timely dataflow computation. It wraps a Allocate, and has a list of child Operates.

Methods

impl<A: Allocate> Root<A>

pub fn new(
c: A,
logging: Arc<Fn(TimelySetup) -> Logger + Sync + Send>
) -> Root<A>

Allocates a new Root bound to a channel allocator.

pub fn step(&mut self) -> bool

Performs one step of the computation.

A step gives each dataflow operator a chance to run, and is the main way to ensure that a computation procedes.

pub fn step_while<F: FnMut() -> bool>(&mut self, func: F)

Calls self.step() as long as func evaluates to true.

pub fn index(&self) -> usize

The index of the worker out of its peers.

pub fn peers(&self) -> usize

The total number of peer workers.

pub fn dataflow<T: Timestamp, R, F: FnOnce(&mut Child<Self, T>) -> R>(
&mut self,
func: F
) -> R

Construct a new dataflow.

pub fn dataflow_using<T: Timestamp, R, F: FnOnce(&mut V, &mut Child<Self, T>) -> R, V: Any + 'static>(
&mut self,
resources: V,
func: F
) -> R

Construct a new dataflow binding resources that are released only after the dataflow is dropped.

This method is designed to allow the dataflow builder to use certain resources that are then stashed with the dataflow until it has completed running. Once complete, the resources are dropped. The most common use of this method at present is with loading shared libraries, where the library is important for building the dataflow, and must be kept around until after the dataflow has completed operation.