Module timely::execute

Starts a timely dataflow execution from configuration information and per-worker logic.

Functions

example
Executes a single-threaded timely dataflow computation.

execute
Executes a timely dataflow from a configuration and per-communicator logic.

execute_from_args
Executes a timely dataflow from supplied arguments and per-communicator logic.