Module timely::logging

Traits, implementations, and macros related to logging timely events.

Structs

ApplicationEvent
Application-defined code start or stop

BatchLogger
Logs events as a timely stream, with progress statements.

ChannelsEvent
The creation of a channel between operators.

CommChannelsEvent
Event on a communication channel

GuardedMessageEvent
Application-defined code start or stop

GuardedProgressEvent
Application-defined code start or stop

InputEvent
Input logic start/stop

MessagesEvent
Message send or receive event

OperatesEvent
The creation of an Operate implementor.

ProgressEvent
Send or receive of progress information.

PushProgressEvent
External progress pushed onto an operator

ScheduleEvent
Operator start or stop.

TimelySetup
Identifier of the worker that generated a log line

Enums

CommChannelKind
Kind of communication channel

StartStop
Records the starting and stopping of an operator.

TimelyEvent
An event in a timely worker

Type Definitions

Logger
Logger type for worker-local logging.

TimelyLogger
Logger for timely dataflow system events.

WorkerIdentifier
Type alias for logging timely events.