Struct timely::logging::ProgressEvent

pub struct ProgressEvent {
    pub is_send: bool,
    pub addr: Vec<usize>,
    pub messages: Vec<(usize, usize, String, i64)>,
    pub internal: Vec<(usize, usize, String, i64)>,
}

Send or receive of progress information.

Fields

is_send

true if the event is a send, and false if it is a receive.

addr

Sequence of nested scope identifiers indicating the path from the root to this instance.

messages

List of message updates, containing Target descriptor, timestamp as string, and delta.

internal

List of capability updates, containing Source descriptor, timestamp as string, and delta.