Module timely::progress

Progress tracking mechanisms to support notification in timely dataflow

Reexports

pub use self::operate::Operate;

pub use self::nested::{Subgraph, Source, Target};

pub use self::timestamp::{Timestamp, PathSummary};

pub use self::count_map::CountMap;

pub use self::frontier::Antichain;

Modules

broadcast
Broadcasts progress information among workers.

count_map
A mapping from general types T to i64, with zero values absent.

frontier
Tracks minimal sets of mutually incomparable elements of a partial order.

nested
Coordination of progress information between a scope-as-operator and its children operators.

notificator
A Notificator manages outstanding capabilities on several inputs as well as notification requests that may be blocked by them.

operate
Methods which describe an operators topology, and the progress it makes.

timestamp
A partially ordered measure of progress at each timely dataflow location.