Struct timely::progress::broadcast::Progcaster

pub struct Progcaster<T: Timestamp> {
    // some fields omitted
}

Manages broadcasting of progress updates to and receiving updates from workers.

Methods

impl<T: Timestamp + Send> Progcaster<T>

fn new<A: Allocate>(allocator: &mut A) -> Progcaster<T>

Creates a new Progcaster using a channel from the supplied allocator.

fn send_and_recv(&mut self, messages: &mut CountMap<(usize, usize, T)>, internal: &mut CountMap<(usize, usize, T)>)

Sends and receives progress updates, broadcasting the contents of messages and internal, and updating each with updates from other workers.