Struct timely::progress::frontier::Antichain

pub struct Antichain<T> { /* fields omitted */ }

A set of mutually incomparable elements.

An antichain is a set of partially ordered elements, each of which is incomparable to the others. This antichain implementation allows you to repeatedly introduce elements to the antichain, and which will evict larger elements to maintain the minimal antichain, those incomparable elements no greater than any other element.

Methods

impl<T: PartialOrder> Antichain<T>

pub fn insert(&mut self, element: T) -> bool

Updates the Antichain if the element is not greater than or equal to some present element.

Returns true if element is added to the set

pub fn new() -> Antichain<T>

Creates a new empty Antichain.

pub fn from_elem(element: T) -> Antichain<T>

Creates a new singleton Antichain.

pub fn clear(&mut self)

Clears the contents of the antichain.

pub fn sort(&mut self) where
T: Ord,

Sorts the elements so that comparisons between antichains can be made.

pub fn less_than(&self, time: &T) -> bool

Returns true if any item in the antichain is strictly less than the argument.

pub fn less_equal(&self, time: &T) -> bool

Returns true if any item in the antichain is less than or equal to the argument.

pub fn dominates(&self, other: &Antichain<T>) -> bool

Returns true if every element of other is greater or equal to some element of self.

Important traits for &'a [u8]
impl<'a> Read for &'a [u8]impl<'a> Write for &'a mut [u8]

pub fn elements(&self) -> &[T]

Reveals the elements in the antichain.