Struct timely::progress::frontier::AntichainRef

pub struct AntichainRef<'a, T: 'a + PartialOrder> { /* fields omitted */ }

A wrapper for elements of an antichain.

Methods

impl<'a, T: 'a + PartialOrder> AntichainRef<'a, T>

pub fn new(frontier: &'a [T]) -> Self

Create a new AntichainRef from a reference to a slice of elements forming the frontier.

pub fn is_empty(&self) -> bool

Returns true if there are no elements in the AntichainRef.

Examples

 use timely::progress::frontier::AntichainRef;

 let frontier = AntichainRef::<usize>::new(&[]);
 assert!(frontier.is_empty());

pub fn iter(&self) -> Iter<T>

Create an iterator over the elements in this AntichainRef.

Examples

 use timely::progress::frontier::AntichainRef;

 let frontier = AntichainRef::new(&[1u64]);
 let mut iter = frontier.iter();
 assert_eq!(iter.next(), Some(&1u64));
 assert_eq!(iter.next(), None);

pub fn less_than(&self, time: &T) -> bool

Returns true if any item in the AntichainRef is strictly less than the argument.

Examples

 use timely::progress::frontier::AntichainRef;

 let frontier = AntichainRef::new(&[1u64]);
 assert!(!frontier.less_than(&0));
 assert!(!frontier.less_than(&1));
 assert!(frontier.less_than(&2));

pub fn less_equal(&self, time: &T) -> bool

Returns true if any item in the AntichainRef is less than or equal to the argument.

Examples

 use timely::progress::frontier::AntichainRef;

 let frontier = AntichainRef::new(&[1u64]);
 assert!(!frontier.less_equal(&0));
 assert!(frontier.less_equal(&1));
 assert!(frontier.less_equal(&2));

pub fn len(&self) -> usize

Returns the number of elements in this AntichainRef.

pub fn to_vec(&self) -> Vec<T> where
T: Clone,

Copies self into a new Vec.