Struct timely::progress::frontier::MutableAntichain

pub struct MutableAntichain<T: Eq> { /* fields omitted */ }

An antichain based on a multiset whose elements frequencies can be updated.

In particular, unlike Antichain, element frequencies can be decremented, removing elements in the antichain and revealing other elements that were previously obscured. As part of this, the MutableAntichain must maintain the frequences for all elements, not just those in the antichain at the moment.

Methods

impl<T: PartialOrder + Eq + Clone + 'static> MutableAntichain<T>

fn new() -> MutableAntichain<T>

Creates a new empty MutableAntichain.

fn clear(&mut self)

Removes all elements from the antichain.

fn elements(&self) -> &[T]

Reveals the element in the MutableAntichain.

fn new_bottom(bottom: T) -> MutableAntichain<T>

Creates a new singleton MutableAntichain.

fn empty(&self) -> bool

Returns true if there are no elements in the MutableAntichain.

fn less_than(&self, time: &T) -> bool

Returns true if any item in the MutableAntichain is strictly less than the argument.

fn less_equal(&self, time: &T) -> bool

Returns true if any item in the MutableAntichain is less than or equal to the argument.

fn count(&self, time: &T) -> Option<i64>

Returns the number of times an element exists in the set.

fn update_into_cm(&mut self, updates: &CountMap<T>, results: &mut CountMap<T>)

Incorporates updates into the antichain, pushing frontier changes into results.

fn update_weight(&mut self, elem: &T, delta: i64, results: &mut CountMap<T>)

Performs a single update to the antichain, pushing frontier changes into results.

fn update(&mut self, elem: &T, delta: i64)

Applies a single update to the antichain.

fn update_iter_and<I: Iterator<Item=(T, i64)>, A: FnMut(&T, i64) -> ()>(&mut self,
updates: I,
action: A)

Applies updates to the antichain and applies action to each frontier change.

fn test_size(&self, threshold: usize, name: &str)

Tests the size of the antichain against a threshould.

This is a diagnostic method in place to observe when antichains become surprisingly large. If either self.occurrences or self.precedents exceeds the threshold, they are listed.

fn update_and<A: FnMut(&T, i64) -> ()>(&mut self,
elem: &T,
delta: i64,
action: A)

Applies an update to the antichain and takes an action on any frontier changes.