Module timely::progress::nested

Coordination of progress information between a scope-as-operator and its children operators.

Reexports

pub use self::subgraph::Subgraph;

pub use self::subgraph::{Source, Target};

pub use self::summary::Summary;

Modules

pointstamp_counter
Manages pointstamp counts (timestamp, location) within a sub operator.

product
A pair timestamp suitable for use with the product partial order.

scope_wrapper
Wraps a child operator, and mediates the communication of progress information.

subgraph
Implements Operate for a scoped collection of child operators.

summary
Path summaries that are either child local, or leave the scope and re-enter from the parent.