Struct timely::progress::nested::pointstamp_counter::PointstampCounter

pub struct PointstampCounter<T: Timestamp> {
    pub source: Vec<Vec<ChangeBatch<T>>>,
    pub target: Vec<Vec<ChangeBatch<T>>>,
    pub pushed: Vec<Vec<ChangeBatch<T>>>,
}

Represents changes to pointstamps before and after transmission along a scope's topology.

Fields
source: Vec<Vec<ChangeBatch<T>>>

timestamp updates indexed by (scope, output)

target: Vec<Vec<ChangeBatch<T>>>

timestamp updates indexed by (scope, input)

pushed: Vec<Vec<ChangeBatch<T>>>

pushed updates indexed by (scope, input)

Methods

impl<T: Timestamp> PointstampCounter<T>

pub fn update_target(&mut self, target: Target, time: T, value: i64)

Updates the count for a time at a target.

pub fn update_source(&mut self, source: Source, time: T, value: i64)

Updates the count for a time at a source.

pub fn clear(&mut self)

Clears the pointstamp counter.

pub fn allocate_for_operator(&mut self, inputs: usize, outputs: usize)

Allocates internal state given an operator's inputs and outputs.