Struct timely::progress::nested::reachability_neu::Tracker

pub struct Tracker<T: Timestamp> { /* fields omitted */ }

An interactive tracker of propagated reachability information.

A Tracker tracks, for a fixed graph topology, the consequences of pointstamp changes at various node input and output ports. These changes may alter the potential pointstamps that could arrive at downstream input ports.

Methods

impl<T: Timestamp> Tracker<T>

pub fn update_target(&mut self, target: Target, time: T, value: i64)

Updates the count for a time at a target.

pub fn update_source(&mut self, source: Source, time: T, value: i64)

Updates the count for a time at a source.

pub fn allocate_from(builder: &Builder<T>) -> Self

Allocate a new Tracker using the shape from summaries.

pub fn propagate_all(&mut self)

Propagates all pending updates.

pub fn pushed(&mut self) -> &mut ChangeBatch<(Target, T)>

A mutable reference to the pushed results of changes.