Struct timely::progress::nested::subgraph::SubgraphBuilder

pub struct SubgraphBuilder<TOuter: Timestamp, TInner: Timestamp> {
    pub name: String,
    pub path: Vec<usize>,
    // some fields omitted
}

A builder structure for initializing Subgraphs.

This collects all the information necessary to get a Subgraph up and running, and is important largely through its build method which actually creates a Subgraph.

Fields
name: String

The name of this subgraph.

path: Vec<usize>

A sequence of integers uniquely identifying the subgraph.

Methods

impl<TOuter: Timestamp, TInner: Timestamp> SubgraphBuilder<TOuter, TInner>

pub fn new_input(
&mut self,
shared_counts: Rc<RefCell<ChangeBatch<Product<TOuter, TInner>>>>
) -> Target

Allocates a new input to the subgraph and returns the target to that input in the outer graph.

pub fn new_output(&mut self) -> Source

Allocates a new output from the subgraph and returns the source of that output in the outer graph.

pub fn connect(&mut self, source: Source, target: Target)

Introduces a dependence from the source to the target.

This method does not effect data movement, but rather reveals to the progress tracking infrastructure that messages produced by source should be expected to be consumed at target.

pub fn new_from(
index: usize,
path: Vec<usize>,
logging: Logger
) -> SubgraphBuilder<TOuter, TInner>

Creates a new Subgraph from a channel allocator and "descriptive" indices.

pub fn allocate_child_id(&mut self) -> usize

Allocates a new child identifier, for later use.

pub fn add_child(
&mut self,
child: Box<Operate<Product<TOuter, TInner>>>,
index: usize,
identifier: usize
)

Adds a new child to the subgraph.

pub fn build<A: Allocate>(self, allocator: &mut A) -> Subgraph<TOuter, TInner>

Now that initialization is complete, actually build a subgraph.