Enum timely::progress::nested::summary::Summary

pub enum Summary<S, T> {
    Local(T),
    Outer(S, T),
}

Summarizes a path within a scope.

The path summary can either be entirely local to the scope (Local) or it may leave the scope and return (Outer).

Variants
Local(T)

Reachable within the scope, and a summary adjusting only the inner coordinate.

Outer(S, T)

Reachable outside the scope, with adjustments to both coordinates.