Trait timely::Push

pub trait Push<T> {
    fn push(&mut self, element: &mut Option<T>);

    fn send(&mut self, element: T) { ... }

    fn done(&mut self) { ... }
}

Pushing elements of type T.

Required Methods

fn push(&mut self, element: &mut Option<T>)

Pushes element and provides the opportunity to take ownership.

The value of element after the call may be changed. A change does not imply anything other than that the implementor took resources associated with element and is returning other resources.

Provided Methods

fn send(&mut self, element: T)

Pushes element and drops any resulting resources.

fn done(&mut self)

Pushes None, conventionally signalling a flush.