Module timely::dataflow

Abstractions for timely dataflow programming.

Timely dataflow programs are constructed by manipulating Stream objects, most often using pre-defined operators that implement known patterns.

Examples

use timely::dataflow::operators::{ToStream, Inspect};

timely::example(|scope| {
    (0..10).to_stream(scope)
           .inspect(|x| println!("seen: {:?}", x));
});

Re-exports

pub use self::stream::Stream;

pub use self::scopes::Scope;

pub use self::scopes::ScopeParent;

pub use self::operators::input::Handle as InputHandle;

pub use self::operators::probe::Handle as ProbeHandle;

Modules

channels
Structured communication between timely dataflow operators.

operators
Extension traits for Stream implementing various operators.

scopes
Hierarchical organization of timely dataflow graphs.

stream
A handle to a typed stream of timely data.