Module timely::dataflow::operators

Extension traits for Stream implementing various operators.

A collection of functions taking typed Stream objects as input and producing new Stream objects as output. Many of the operators provide simple, composable functionality. Some of the operators are more complicated, for use with advanced timely dataflow features.

The Unary and Binary operators provide general operators whose behavior can be supplied using closures accepting input and output handles. Most of the operators in this module are defined using these two general operators.

Reexports

pub use self::enterleave::{Enter, EnterAt, Leave};

pub use self::unary::Unary;

pub use self::queue::*;

pub use self::input::Input;

pub use self::feedback::{LoopVariable, ConnectLoop};

pub use self::concat::{Concat, Concatenate};

pub use self::partition::Partition;

pub use self::map::Map;

pub use self::inspect::Inspect;

pub use self::filter::Filter;

pub use self::binary::Binary;

pub use self::delay::Delay;

pub use self::exchange::Exchange as ExchangeExtension;

pub use self::probe::Probe;

pub use self::to_stream::ToStream;

pub use self::capture::Capture;

Modules

binary
Methods to construct generic streaming and blocking binary operators.

capture
An operator which captures the streams of records and notifications to play back later.

concat

delay

enterleave
Extension traits to move a Stream between an outer Scope and inner Scope.

exchange
Exchange records between workers.

feedback
Create cycles in a timely dataflow graph.

filter
Filters a stream by a predicate.

input
Create new Streams connected to external inputs.

inspect
Extension trait and implementation for observing and action on streamed data.

map
Extension methods for Stream based on record-by-record transformation.

partition
Partition a stream of records into multiple streams.

probe
Monitor progress at a Stream.

queue

to_stream
Conversion to the Stream type from iterators.

unary
Methods to construct generic streaming and blocking unary operators.