Struct timely::dataflow::operators::input::Handle

pub struct Handle<T: Timestamp + Ord, D: Data> {
    // some fields omitted
}

A handle to an input Stream, used to introduce data to a timely dataflow computation.

Methods

impl<T: Timestamp + Ord, D: Data> Handle<T, D>

fn send(&mut self, data: D)

Sends one record into the corresponding timely dataflow Stream, at the current epoch.

fn advance_to(&mut self, next: T)

Advances the current epoch to next.

This method allows timely dataflow to issue progress notifications as it can now determine that this input can no longer produce data at earlier timestamps.

fn close(self)

Closes the input.

This method allows timely dataflow to issue all progress notifications blocked by this input and to begin to shut down operators, as this input can no longer produce data.

fn epoch(&mut self) -> &T

Reports the current epoch.