Struct timely::dataflow::operators::Capability

pub struct Capability<T: Timestamp> { /* fields omitted */ }

A capability for timestamp t represents a permit for an operator that holds the capability to send data and request notifications at timestamp t.

Methods

impl<T: Timestamp> Capability<T>

fn time(&self) -> T

The timestamp associated with this capability.

fn delayed(&self, new_time: &T) -> Capability<T>

Makes a new capability for a timestamp new_time greater or equal to the timestamp of the source capability (self).