Struct timely::dataflow::scopes::child::Child

pub struct Child<'a, G: ScopeParent, T: Timestamp> {
    pub subgraph: &'a RefCell<SubgraphBuilder<G::Timestamp, T>>,
    pub parent: G,
    pub logging: Logger,
}

A Child wraps a Subgraph and a parent G: Scope. It manages the addition of Operates to a subgraph, and the connection of edges between them.

Fields
subgraph: &'a RefCell<SubgraphBuilder<G::Timestamp, T>>

The subgraph under assembly.

parent: G

A copy of the child's parent scope.

logging: Logger

The log writer for this scope.

Methods

impl<'a, G: ScopeParent, T: Timestamp> Child<'a, G, T>

pub fn index(&self) -> usize

This worker's unique identifier.

Ranges from 0 to self.peers() - 1.

pub fn peers(&self) -> usize

The total number of workers in the computation.