Struct timely::dataflow::scopes::child::Child

pub struct Child<G: Scope, T: Timestamp> {
    pub subgraph: Rc<RefCell<Subgraph<G::Timestamp, T>>>,
    pub parent: G,
}

A Child wraps a Subgraph and a parent G: Scope. It manages the addition of Operates to a subgraph, and the connection of edges between them.

Fields

subgraph

parent