Struct timely::dataflow::scopes::root::Root

pub struct Root<A: Allocate> {
    // some fields omitted
}

A Root is the entry point to a timely dataflow computation. It wraps a Allocate, and has a slot for one child Operate. The primary intended use of Root is through its implementation of the Scope trait.

Methods

impl<A: Allocate> Root<A>

fn new(c: A) -> Root<A>

fn step(&mut self) -> bool

fn index(&self) -> usize

fn peers(&self) -> usize