Struct timely::dataflow::stream::Stream

pub struct Stream<S: Scope, D> { /* fields omitted */ }

Abstraction of a stream of D: Data records timestamped with S::Timestamp.

Internally Stream maintains a list of data recipients who should be presented with data produced by the source of the stream.

Methods

impl<S: Scope, D> Stream<S, D>

pub fn connect_to<P: Push<Bundle<S::Timestamp, D>> + 'static>(
&self,
target: Target,
pusher: P,
identifier: usize
)

Connects the stream to a destination.

The destination is described both by a Target, for progress tracking information, and a P: Push where the records should actually be sent. The identifier is unique to the edge and is used only for logging purposes.

pub fn new(source: Source, output: TeeHelper<S::Timestamp, D>, scope: S) -> Self

Allocates a Stream from a supplied Source name and rendezvous point.

pub fn name(&self) -> &Source

The name of the stream's source operator.

pub fn scope(&self) -> S

The scope immediately containing the stream.