Struct timely::dataflow::stream::Stream

pub struct Stream<S: Scope, D: Data> {
    // some fields omitted
}

Abstraction of a stream of D: Data records timestamped with S::Timestamp.

Internally Stream maintains a list of data recipients who should be presented with data produced by the source of the stream.

Methods

impl<S: Scope, D: Data> Stream<S, D>

fn connect_to<P: Push<(S::Timestamp, Content<D>)> + 'static>(&self, target: Target, pusher: P, identifier: usize)

Connects the stream to a destination.

The destination is described both by a Target, for progress tracking information, and a P: Push where the records should actually be sent. The identifier is unique to the edge and is used only for logging purposes.

fn new(source: Source, output: TeeHelper<S::Timestamp, D>, scope: S) -> Self

Allocates a Stream from a supplied Source name and rendezvous point.

fn name(&self) -> &Source

The name of the stream's source operator.

fn scope(&self) -> S

The scope immediately containing the stream.