Struct timely::dataflow::stream::Stream

pub struct Stream<S: Scope, D: Data> {
    // some fields omitted
}

Abstraction of a stream of D: Data records timestamped with S::Timestamp.

Internally Stream maintains a list of data recipients who should be presented with data produced by the source of the stream.

Methods

impl<S: Scope, D: Data> Stream<S, D>

fn connect_to<P: Push<(S::Timestamp, Content<D>)> + 'static>(&self, target: Target, pusher: P, _identifier: usize)

fn new(source: Source, output: TeeHelper<S::Timestamp, D>, scope: S) -> Self

fn name(&self) -> &Source

fn scope(&self) -> S