Function timely::execute::example

pub fn example<F: Fn(&mut Child<Root<Allocator>, u64>) + Send + Sync + 'static>(func: F)

Executes a single-threaded timely dataflow computation.

The example method takes a closure on a Scope which it executes to initialize and run a timely dataflow computation on a single thread. This method is intended for use in examples, rather than programs that may need to run across multiple workers.

Examples

use timely::dataflow::operators::{ToStream, Inspect};

timely::example(|scope| {
    (0..10).to_stream(scope)
           .inspect(|x| println!("seen: {:?}", x));
});