Function timely::execute::execute

pub fn execute<F: Fn(&mut Root<Allocator>) + Send + Sync + 'static>(config: Configuration, func: F)

Executes a timely dataflow from a configuration and per-communicator logic.

The execute method takes a Configuration and spins up some number of workers threads, each of which execute the supplied closure to construct and run a timely dataflow computation.

Examples

use timely::dataflow::Scope;
use timely::dataflow::operators::{ToStream, Inspect};

// execute a timely dataflow using three worker threads.
timely::execute(timely::Configuration::Process(3), |root| {
    root.scoped::<u64,_,_>(|scope| {
        (0..10).to_stream(scope)
               .inspect(|x| println!("seen: {:?}", x));
    })
});