Struct timely::logging::EventPusherTee

pub struct EventPusherTee<T: PartialOrder + Ord + Default + Clone + 'static, D: Clone> { /* fields omitted */ }

An EventPusher that supports dynamically adding new EventPushers.

The tee maintains the frontier as the stream of events passes by. When a new pusher arrives it advances the frontier to the current value, and starts to forward events to it as well.

Methods

impl<T: PartialOrder + Ord + Default + Clone + 'static, D: Clone> EventPusherTee<T, D>

pub fn new() -> Self

Construct a new tee with no subscribers.

pub fn subscribe(&mut self, listener: Box<EventPusher<T, D> + Send>)

Subscribe to this tee.