Trait timely::order::PartialOrder

pub trait PartialOrder: Eq {
    fn less_equal(&self, other: &Self) -> bool;

    fn less_than(&self, other: &Self) -> bool { ... }
}

A type that is partially ordered.

This trait is distinct from Rust's PartialOrd trait, because the implementation of that trait precludes a distinct Ord implementation. We need an independent trait if we want to have a partially ordered type that can also be sorted.

Required methods

fn less_equal(&self, other: &Self) -> bool

Returns true iff one element is less than or equal to the other.

Loading content...
Provided methods

fn less_than(&self, other: &Self) -> bool

Returns true iff one element is strictly less than the other.

Loading content...