Struct timely::progress::frontier::Antichain

pub struct Antichain<T> {
    // some fields omitted
}

A set of mutually incomparable elements.

Methods

impl<T: PartialOrd + Eq + Copy + Debug> Antichain<T>

fn insert(&mut self, element: T) -> bool

Updates the Antichain if the element is not greater than or equal to some present element.

Returns true if element is added to the set

fn new() -> Antichain<T>

Creates a new empty Antichain.

fn from_elem(element: T) -> Antichain<T>

Creates a new singleton Antichain.

fn elements(&self) -> &[T]

Reveals the elements in the Antichain.