Struct timely::progress::frontier::MutableAntichain

pub struct MutableAntichain<T: Eq> {
    // some fields omitted
}

An antichain based on a multiset whose elements frequencies can be updated.

In particular, unlike Antichain, element frequencies can be decremented, removing elements in the antichain and revealing other elements that were previously obscured. As part of this, the MutableAntichain must maintain the frequences for all elements, not just those in the antichain at the moment.

Methods

impl<T: PartialOrd + Eq + Clone + Debug + 'static> MutableAntichain<T>

fn new() -> MutableAntichain<T>

Creates a new empty MutableAntichain.

fn elements(&self) -> &[T]

Reveals the element in the MutableAntichain.

fn new_bottom(bottom: T) -> MutableAntichain<T>

Creates a new singleton MutableAntichain.

fn empty(&self) -> bool

Returns true if there are no elements in the MutableAntichain.

fn lt(&self, time: &T) -> bool

Returns true if any item in the MutableAntichain is strictly less than the argument.

fn le(&self, time: &T) -> bool

Returns true if any item in the MutableAntichain is less than or equal to the argument.

fn count(&self, time: &T) -> Option<i64>

Returns the number of times an element exists in the set.

fn update_into_cm(&mut self, updates: &CountMap<T>, results: &mut CountMap<T>)

fn update_weight(&mut self, elem: &T, delta: i64, results: &mut CountMap<T>)

fn update(&mut self, elem: &T, delta: i64)

fn update_iter_and<I: Iterator<Item=(T, i64)>, A: FnMut(&T, i64) -> ()>(&mut self, updates: I, action: A)

fn test_size(&self, threshold: usize, name: &str)

fn update_and<A: FnMut(&T, i64) -> ()>(&mut self, elem: &T, delta: i64, action: A)