Struct timely::progress::nested::product::Product

pub struct Product<TOuter, TInner> {
    pub outer: TOuter,
    pub inner: TInner,
}

A nested pair of timestamps, one outer and one inner.

We use Product rather than (TOuter, TInner) so that we can derive our own PartialOrd, because Rust just uses the lexicographic total order.

Fields
outer: TOuter

Outer timestamp.

inner: TInner

Inner timestamp.

Methods

impl<TOuter, TInner> Product<TOuter, TInner>

pub fn new(outer: TOuter, inner: TInner) -> Product<TOuter, TInner>

Creates a new product from outer and inner coordinates.