Struct timely::progress::nested::subgraph::Subgraph

pub struct Subgraph<TOuter: Timestamp, TInner: Timestamp> {
    pub path: Vec<usize>,
    // some fields omitted
}

A dataflow subgraph.

The subgraph type contains the infrastructure required to describe the topology of and track progress within a dataflow subgraph.

Fields
path: Vec<usize>

Path of identifiers from the root.