Struct timely::progress::nested::subgraph::Subgraph

pub struct Subgraph<TOuter: Timestamp, TInner: Timestamp> {
    pub path: Vec<usize>,
    pub index: usize,
    // some fields omitted
}

A dataflow subgraph.

The subgraph type contains the infrastructure required to describe the topology of and track progress within a dataflow subgraph.

Fields
path: Vec<usize>

A sequence of integers uniquely identifying the subgraph.

index: usize

The index assigned to the subgraph by its parent.

Methods

impl<TOuter: Timestamp, TInner: Timestamp> Subgraph<TOuter, TInner>

fn new_input(&mut self,
shared_counts: Rc<RefCell<CountMap<Product<TOuter, TInner>>>>)
-> usize

Allocates a new input to the subgraph and returns the assigned index.

fn new_output(&mut self) -> usize

Allocates a new output from the subgraph and returns the assigned index.

fn connect(&mut self, source: Source, target: Target)

Introduces a dependence from the source to the target.

This method does not effect data movement, but rather reveals to the progress tracking infrastructure that messages produced by source should be expected to be consumed at target.

fn new_from<A: Allocate>(allocator: &mut A,
index: usize,
path: Vec<usize>)
-> Subgraph<TOuter, TInner>

Creates a new Subgraph from a channel allocator and "descriptive" indices.

fn allocate_child_id(&mut self) -> usize

Allocates a new child identifier, for later use.

fn add_child(&mut self,
child: Box<Operate<Product<TOuter, TInner>>>,
index: usize,
identifier: usize)

Adds a new child to the subgraph.