Module timely::progress::timestamp

A partially ordered measure of progress at each timely dataflow location.

Structs

RootSummary
An empty path summary for root timestamps.

RootTimestamp
An empty timestamp used by the root scope.

Traits

PathSummary
A summary of how a timestamp advances along a timely dataflow path.

Timestamp
A composite trait for types that serve as timestamps in timely dataflow.