Trait timely::progress::timestamp::PathSummary

pub trait PathSummary<T>: 'static + Copy + Eq + PartialOrd + Debug + Default {
    fn results_in(&self, src: &T) -> T;
    fn followed_by(&self, other: &Self) -> Self;
}

A summary of how a timestamp advances along a timely dataflow path.

Required Methods

fn results_in(&self, src: &T) -> T

Advances a timestamp according to the timestamp actions on the path.

fn followed_by(&self, other: &Self) -> Self

Composes this path summary with another path summary.