Trait timely::progress::timestamp::Timestamp

pub trait Timestamp: Clone + Eq + PartialOrder + Default + Debug + Send + Any + Abomonation + Hash + Ord {
    type Summary: PathSummary<Self> + 'static;
}

A composite trait for types that serve as timestamps in timely dataflow.

Associated Types

type Summary: PathSummary<Self> + 'static

A type summarizing action on a timestamp along a dataflow path.